refac: navigation declarations
