+++ Globoplay/Sources/Views/ContentView.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import SwiftData
 
 struct ContentView: View {
 	init() {
@@ -17,9 +16,7 @@ struct ContentView: View {
 		TabView {
 			NavigationStack {
 				HomeView()
-					.navigationDestination(for: ContentViewData.self) { viewData in
-						ContentDetailsView(viewData: viewData)
-					}
+					.routeToContentViewDetails()
 			}
 			.tabItem {
 				Label("Início", systemImage: "house.fill")
@@ -28,48 +25,12 @@ struct ContentView: View {
 
 			NavigationStack {
 				MyListView()
-					.navigationDestination(for: ContentViewData.self) { viewData in
-						ContentDetailsView(viewData: viewData)
-					}
+					.routeToContentViewDetails()
 			}
-
 			.tabItem {
 				Label("Minha lista", systemImage: "star.fill")
 					.font(.caption2)
 			}
-		}
-	}
-}
-
-struct MyListView: View {
-	@Query var contentsSaved: [ContentModel]
-
-	var body: some View {
-		ContentsGridView(viewDatas: contentsSaved.lazy.map {
-			ContentViewData(content: $0)
-		})
-	}
-}
-
-struct ContentsGridView: View {
-	let viewDatas: [ContentViewData]
-
-	let columns = [
-		GridItem(.flexible()),
-		GridItem(.flexible()),
-		GridItem(.flexible()),
-	]
-
-	var body: some View {
-		ScrollView {
-			LazyVGrid(columns: columns, spacing: 20) {
-				ForEach(viewDatas, id: \.content.id) { viewData in
-					NavigationLink(value: viewData) {
-						ContentImageView(viewData: viewData, size: .small)
-					}
-				}
-			}
-			.padding(.horizontal)
 		}
 	}
 }
